remove unused native libs
These files are from the zip file Discord provides of all native
libraries, and are probably useful for supporting some OSes. However,
the current code will never load anything other than a `.dll` or `.so`,
so they're just taking up space in the JAR. They can be re-added if
needed.